Implementar filtro potencia.

+++ app/src/main/java/edu/iesam/bikerly/data/local/mock/Mappers.kt
@@ -10,7 +10,7 @@ fun MotorbikeMockModel.toModel(): Motorbike {
         this.model,
         this.year,
         this.type,
-        this.displacement,
+        this.displacement.toInt(),
         this.img.toUri()
     )
 }
